git-svn-id: svn://svn.riscos.info/gccsdk/trunk@491 5d7976f4-0b0e-0410-a586-d2bac8e3be52

+++ gcc/riscos-aof/as/macros.h
@@ -54,7 +54,7 @@ extern Macro *macroCurrent;
 extern char *macroPtr;
 extern long int macroCurrentCallNo;
 
-Macro *macroFind (int len, char *);
+Macro *macroFind (size_t len, char *);
 void macroCall (Macro *, Lex *);
 BOOL macroGetLine (char *buf);	/* returns 0 if end of macro */
